Separating test docs
