+++ docs/examples/testing/pytest_request_factory.py
@@ -1,0 +1,24 @@
+import json
+from http import HTTPStatus
+
+from dirty_equals import IsUUID
+from django.http import HttpResponse
+
+from dmr.test import DMRRequestFactory
+from examples.testing.pydantic_controller import UserController
+
+
+def test_request_factory(dmr_rf: DMRRequestFactory) -> None:
+    request_data = {'email': 'test@example.com', 'age': 43}
+    request = dmr_rf.post('/users/', data=request_data)
+
+    response = UserController.as_view()(request)
+
+    assert isinstance(response, HttpResponse)
+    assert request.content_type == 'application/json'
+    assert response.status_code == HTTPStatus.CREATED
+    assert response.headers == {'Content-Type': 'application/json'}
+    assert json.loads(response.content) == {
+        'uid': IsUUID,
+        **request_data,
+    }
