+++ docs/index.rst
@@ -207,7 +207,19 @@ Here are our amazing people who made this project possible.
   pages/external-views.rst
   pages/integrations.rst
   pages/internationalization.rst
-  pages/testing.rst
+
+
+.. toctree::
+  :caption: Testing
+  :hidden:
+
+  pages/testing/tools-and-styles.rst
+  pages/testing/authentication.rst
+  pages/testing/throttling.rst
+  pages/testing/data-generation.rst
+  pages/testing/property-based.rst
+  pages/testing/tracecov.rst
+  pages/testing/api-reference.rst
 
 
 .. toctree::
